requirements.txt
asgiref==3.7.1
Django==4.0.10
djangorestframework==3.14.0
psycopg2-binary==2.9.3
sqlparse==0.4.4
typing_extensions==4.6.2
